-FoundObjects There is a FoundObjects enum containing 6 values: Mundane, Bucket_Spade, Holograms, Glasses, Model, Disk Mundane is for any object that there is no puzzle for. We can decide if we want to do handle all mundane objects a certain way. ie. displaying filler text "This is a vase, it looks useless." - but probably as a 'if we have time' type thing
-FoundObject
FoundObject should be added to any game object the player is expected to be able to click on. Assign the type of object it should be, and assign the texture that will be used if it's not a Mundane object. There are rules for how the texture should be set up, read/write need to be enabled under advanced settings and filter type to point, this info also comes up as a tool tip, hover over the texture field.

-ObjectSelecting
This object can be put on the camera, but it doesn't have to be. But it does need to know the game view camera. Drag it into the variable field.